move unwrapNestedDivs to scripts.js
