+++ blocks/contact-us/contact-us.js
@@ -1,4 +1,5 @@
 import { div } from '../../scripts/dom-helpers.js';
+import { unwrapNestedDivs } from '../../scripts/scripts.js';
 
 function changeElementTag(oldElement, newTagName) {
   const newElement = document.createElement(newTagName);
@@ -12,19 +13,6 @@ function changeElementTag(oldElement, newTagName) {
 }
 
 export default async function decorate(block) {
-  // Function to recursively unwrap nested divs
-  function unwrapNestedDivs(element) {
-    const children = Array.from(element.children);
-    children.forEach((child) => {
-      if (child.tagName === 'DIV') {
-        unwrapNestedDivs(child);
-        while (child.firstChild) {
-          element.insertBefore(child.firstChild, child);
-        }
-        element.removeChild(child);
-      }
-    });
-  }
   unwrapNestedDivs(block);
 
   const heading = div({ class: 'heading', tabIndex: 0 });
