+++ blocks/quote/quote.js
@@ -1,19 +1,7 @@
 import { div, h1, h3 } from '../../scripts/dom-helpers.js';
+import { unwrapNestedDivs } from '../../scripts/scripts.js';
 
 export default async function decorate(block) {
-  // Function to recursively unwrap nested divs
-  function unwrapNestedDivs(element) {
-    const children = Array.from(element.children);
-    children.forEach((child) => {
-      if (child.tagName === 'DIV') {
-        unwrapNestedDivs(child);
-        while (child.firstChild) {
-          element.insertBefore(child.firstChild, child);
-        }
-        element.removeChild(child);
-      }
-    });
-  }
   unwrapNestedDivs(block);
 
   const quoteWrapper = div(
